Read or Make RDS

Reads or makes .rds if file is not available and loads the object into memory.

rom(
  obj_name,
  ...,
  dir = "rds",
  file_name_sufix = "",
  file_name_sufix_sep = ".",
  file_name_extention = ".rds",
  obj_name_has_extention = FALSE,
  compress_rds = FALSE,
  return_obj_name = FALSE,
  do_not_make = FALSE,
  quietly = FALSE,
  assign_to_obj_name = TRUE
)

Arguments

obj_name

...

dir

Directory where to look for or save to .rds. Default is "./rds"

file_name_sufix

e.g., date or version. Default is ""

file_name_sufix_sep

Default is "."

file_name_extention

Default is ".rds"

obj_name_has_extention

Default is FALSE

compress_rds

Default is FALSE. It is faster to read and write if rds is not compressed.

return_obj_name

Default is FALSE

do_not_make

Default is FALSE

quietly

Default is FALSE

assign_to_obj_name

Value

Returns the object or name of the obj (if return_obj_name is TRUE).